+++ src/main/java/nextstep/domain/member/MemberDao.java
@@ -37,6 +37,7 @@ public class MemberDao {
                 ps.setString(2, member.getPassword());
                 ps.setString(3, member.getName());
                 ps.setString(4, member.getPhone());
+                ps.setString(5, member.getRole().toString());
                 return ps;
 
             }, keyHolder);
