+++ src/main/java/nextstep/domain/reservation/ReservationDao.java
@@ -41,6 +41,7 @@ public class ReservationDao {
                             resultSet.getString("member.password"),
                             resultSet.getString("member.name"),
                             resultSet.getString("member.phone")
+
                             )
     );
