+++ src/main/java/nextstep/interfaces/member/dto/MemberRequest.java
@@ -10,6 +10,9 @@ public class MemberRequest {
     private String phone;
     private MemberRole role;
 
+    public MemberRequest() {
+    }
+
     public MemberRequest(String username, String password, String name, String phone, MemberRole role) {
         this.username = username;
         this.password = password;
@@ -47,6 +50,6 @@ public class MemberRequest {
     }
 
     public Member toEntity() {
-        return new Member(username, password, name, phone);
+        return new Member(username, password, name, phone, role);
     }
 }
